requirements.txt
ccxt
pandas
matplotlib
requests
Pillow
schedule
tk
